Player next

+++ ff-media-tv-app/src/main/java/com/faithForward/media/viewModel/uiModels/GenreUiModel.kt
@@ -18,7 +18,7 @@ fun ContentItem.toGenreCardDto(): GenreCardDto {
         genre = genres?.mapNotNull { it.name }?.filter { it.isNotBlank() }?.joinToString(", ")
             .takeIf { !it.isNullOrBlank() },
         seasons = seasons?.takeIf { it.isNotEmpty() }?.size,
-        duration = duration?.let { formatDuration(it) } ?: "",
+        duration = duration?.let { formatDuration(it.toLong()) } ?: "",
         imdbRating = rating.takeIf { !it.isNullOrBlank() },
         releaseDate = dateUploaded.takeIf { !it.isNullOrBlank() },
         videoUrl = video_link.takeIf { !it.isNullOrBlank() },
